+++ download_gp_id.py
@@ -5,6 +5,7 @@ import spacetrack.operators as op
 from spacetrack import SpaceTrackClient
 import time
 import sys
+import os
 import math
 import argparse
 import subprocess
@@ -54,14 +55,21 @@ def getdata(st, norad_cat_id, logger = None):
 
 getdata.lasttime = 0
 
-def savedata(data, filename, compress = True, logger = None):
+def savedata(data, filename, compress = True, force = False, logger = None):
+    if os.path.exists(filename) and not force:
+        logger.error(filename + ' already exists')
+        return False
+
     with open(filename, 'w') as fp:
         fp.write(data)
 
     if not compress:
         return True
 
-    proc = subprocess.Popen(['xz', '-9', filename], stdout=PIPE, stderr=PIPE, text=True)
+    program = ['xz', '-9', filename]
+    if force:
+        program.insert(1, '-f')
+    proc = subprocess.Popen(program, stdout=PIPE, stderr=PIPE, text=True)
     (stdout, stderr) = proc.communicate()
     if stdout != '' or stderr != '':
         logger.error(stdout + stderr)
@@ -76,7 +84,9 @@ def main():
     parser.add_argument('START', type=int, help='Start Catalog Number.')
     parser.add_argument('END', type=int, nargs='?', help='End Catalog Number. Default: same as START')
     parser.add_argument('CHUNK', type=int, nargs='?', default=1, help='Chunk size. Default: 1')
+    parser.add_argument('-f', '--force', action='store_true', help='If the outpu file already exists, overwrite it.')
     args = parser.parse_args()
+    force = args.force
 
     start = args.START
     end = args.END if args.END is not None else start
@@ -121,7 +131,7 @@ def main():
         else:
             tsize += len(data)
             tfiles += 1
-            result = savedata(data, filename, logger = logger)
+            result = savedata(data, filename, force = force, logger = logger)
             if not result:
                 logger.error("Error: Fail to save data for NORAD Catalog Number {}".format(norad_cat_id))
                 error_count += 1
